rename pyramid.config.derivations to pyramid.viewderivers
