+++ pyramid/tests/test_viewderivers.py
@@ -46,7 +46,7 @@ class TestDeriveView(unittest.TestCase):
             self.assertEqual(
                 e.args[0],
                 'Could not convert return value of the view callable function '
-                'pyramid.tests.test_config.test_derivations.view into a response '
+                'pyramid.tests.test_viewderivers.view into a response '
                 'object. The value returned was None. You may have forgotten '
                 'to return a value from the view callable.'
                 )
@@ -64,7 +64,7 @@ class TestDeriveView(unittest.TestCase):
             self.assertEqual(
                 e.args[0],
                 "Could not convert return value of the view callable function "
-                "pyramid.tests.test_config.test_derivations.view into a response "
+                "pyramid.tests.test_viewderivers.view into a response "
                 "object. The value returned was {'a': 1}. You may have "
                 "forgotten to define a renderer in the view configuration."
                 )
@@ -84,7 +84,7 @@ class TestDeriveView(unittest.TestCase):
             msg = e.args[0]
             self.assertTrue(msg.startswith(
                 'Could not convert return value of the view callable object '
-                '<pyramid.tests.test_config.test_derivations.'))
+                '<pyramid.tests.test_viewderivers.'))
             self.assertTrue(msg.endswith(
                 '> into a response object. The value returned was None. You '
                 'may have forgotten to return a value from the view callable.'))
@@ -128,7 +128,7 @@ class TestDeriveView(unittest.TestCase):
                 e.args[0],
                 'Could not convert return value of the view callable '
                 'method __call__ of '
-                'class pyramid.tests.test_config.test_derivations.AView into a '
+                'class pyramid.tests.test_viewderivers.AView into a '
                 'response object. The value returned was None. You may have '
                 'forgotten to return a value from the view callable.'
                 )
@@ -151,7 +151,7 @@ class TestDeriveView(unittest.TestCase):
                 e.args[0],
                 'Could not convert return value of the view callable '
                 'method theviewmethod of '
-                'class pyramid.tests.test_config.test_derivations.AView into a '
+                'class pyramid.tests.test_viewderivers.AView into a '
                 'response object. The value returned was None. You may have '
                 'forgotten to return a value from the view callable.'
                 )
@@ -358,7 +358,7 @@ class TestDeriveView(unittest.TestCase):
         self.assertFalse(result is view)
         self.assertEqual(view.__module__, result.__module__)
         self.assertEqual(view.__doc__, result.__doc__)
-        self.assertTrue('test_derivations' in result.__name__)
+        self.assertTrue('test_viewderivers' in result.__name__)
         self.assertFalse(hasattr(result, '__call_permissive__'))
         self.assertEqual(result(None, None), response)
